convert float to double

+++ test/nl/tue/c2IOE0/group5/ai/data/GameSimulatorTest.java
@@ -1,22 +1,15 @@
 package nl.tue.c2IOE0.group5.ai.data;
 
 import nl.tue.c2IOE0.group5.ai.GameSimulator;
-import nl.tue.c2IOE0.group5.controllers.PlayerController;
-import nl.tue.c2IOE0.group5.enemies.Enemy;
 import nl.tue.c2IOE0.group5.enemies.EnemyType;
-import nl.tue.c2IOE0.group5.engine.Engine;
-import nl.tue.c2IOE0.group5.engine.Simulator;
-import nl.tue.c2IOE0.group5.providers.AnimationProvider;
-import nl.tue.c2IOE0.group5.providers.BackgroundProvider;
 import nl.tue.c2IOE0.group5.providers.EnemyProvider;
 import nl.tue.c2IOE0.group5.providers.TowerProvider;
 import nl.tue.c2IOE0.group5.towers.TowerType;
-import org.joml.Vector2i;
 import org.junit.Test;
 
 import java.io.IOException;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertTrue;
 
 public class GameSimulatorTest {
     @Test
@@ -52,7 +45,7 @@ public class GameSimulatorTest {
 
         sim.run();
 
-        float score = sim.getDestructionScore();
+        double score = sim.getDestructionScore();
         System.out.println("Score: " + score);
         assertTrue(score > 0 && score <= 1);
     }
